add CommandPlaintext
