+++ modules/anonify-enclave/src/commands.rs
@@ -10,12 +10,16 @@ use frame_enclave::EnclaveEngine;
 use frame_runtime::traits::*;
 use frame_sodium::SodiumCiphertext;
 use serde::{Deserialize, Serialize};
-use std::{marker::PhantomData, vec::Vec};
+use std::{
+    marker::PhantomData,
+    string::{String, ToString},
+    vec::Vec,
+};
 
 /// A message sender that encrypts commands
 #[derive(Debug, Clone, Default)]
 pub struct CmdSender<AP: AccessPolicy> {
-    ecall_input: input::Command<AP>,
+    command_plaintext: CommandPlaintext<AP>,
 }
 
 impl<AP> EnclaveEngine for CmdSender<AP>
@@ -30,12 +34,12 @@ where
         C: ContextOps<S = StateType> + Clone,
     {
         let buf = enclave_context.decrypt(ciphertext)?;
-        let ecall_input = serde_json::from_slice(&buf[..])?;
-        Ok(Self { ecall_input })
+        let command_plaintext = serde_json::from_slice(&buf[..])?;
+        Ok(Self { command_plaintext })
     }
 
     fn eval_policy(&self) -> anyhow::Result<()> {
-        self.ecall_input.access_policy().verify()
+        self.command_plaintext.access_policy().verify()
     }
 
     fn handle<R, C>(self, enclave_context: &C, max_mem_size: usize) -> anyhow::Result<Self::EO>
@@ -48,8 +52,8 @@ where
         // ratchet sender's app keychain per tx.
         group_key.sender_ratchet(roster_idx as usize)?;
 
-        let my_account_id = self.ecall_input.access_policy().into_account_id();
-        let ciphertext = Commands::<R, C, AP>::new(my_account_id, self.ecall_input)?
+        let my_account_id = self.command_plaintext.access_policy().into_account_id();
+        let ciphertext = CommandExecutor::<R, C, AP>::new(my_account_id, self.command_plaintext)?
             .encrypt(group_key, max_mem_size)?;
 
         let msg = Sha256::hash_for_attested_tx(
@@ -62,6 +66,60 @@ where
         let command_output = output::Command::new(ciphertext, enclave_sig.0, enclave_sig.1);
 
         Ok(command_output)
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct CommandPlaintext<AP: AccessPolicy> {
+    #[serde(deserialize_with = "AP::deserialize")]
+    pub access_policy: AP,
+    pub runtime_params: serde_json::Value,
+    pub cmd_name: String,
+    pub counter: UserCounter,
+}
+
+impl<AP> Default for CommandPlaintext<AP>
+where
+    AP: AccessPolicy,
+{
+    fn default() -> Self {
+        Self {
+            access_policy: AP::default(),
+            runtime_params: serde_json::Value::Null,
+            cmd_name: String::default(),
+            counter: UserCounter::default(),
+        }
+    }
+}
+
+impl<AP> CommandPlaintext<AP>
+where
+    AP: AccessPolicy,
+{
+    pub fn new(
+        access_policy: AP,
+        runtime_params: serde_json::Value,
+        cmd_name: impl ToString,
+        counter: UserCounter,
+    ) -> Self {
+        CommandPlaintext {
+            access_policy,
+            runtime_params,
+            cmd_name: cmd_name.to_string(),
+            counter,
+        }
+    }
+
+    pub fn access_policy(&self) -> &AP {
+        &self.access_policy
+    }
+
+    pub fn cmd_name(&self) -> &str {
+        &self.cmd_name
+    }
+
+    pub fn counter(&self) -> UserCounter {
+        self.counter
     }
 }
 
@@ -120,7 +178,7 @@ where
 
         let mut output = output::ReturnNotifyState::default();
         let decrypted_cmds =
-            Commands::<R, C, AP>::decrypt(self.ecall_input.ciphertext(), group_key)?;
+            CommandExecutor::<R, C, AP>::decrypt(self.ecall_input.ciphertext(), group_key)?;
         if let Some(cmds) = decrypted_cmds {
             // Since the command data is valid for the error at the time of state transition,
             // `user_counter` must be verified and incremented before the state transition.
@@ -141,7 +199,7 @@ where
 
 /// Command data which make state update
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Commands<R: RuntimeExecutor<CTX>, CTX: ContextOps<S = StateType>, AP> {
+pub struct CommandExecutor<R: RuntimeExecutor<CTX>, CTX: ContextOps<S = StateType>, AP> {
     my_account_id: AccountId,
     #[serde(deserialize_with = "R::C::deserialize")]
     call_kind: R::C,
@@ -150,19 +208,22 @@ pub struct Commands<R: RuntimeExecutor<CTX>, CTX: ContextOps<S = StateType>, AP>
     ap: PhantomData<AP>,
 }
 
-impl<R, CTX, AP> Commands<R, CTX, AP>
+impl<R, CTX, AP> CommandExecutor<R, CTX, AP>
 where
     R: RuntimeExecutor<CTX, S = StateType>,
     CTX: ContextOps<S = StateType>,
     AP: AccessPolicy,
 {
-    pub fn new(my_account_id: AccountId, ecall_input: input::Command<AP>) -> Result<Self> {
-        let call_kind = R::C::new(ecall_input.cmd_name(), ecall_input.runtime_params.clone())?;
+    pub fn new(my_account_id: AccountId, command_plaintext: CommandPlaintext<AP>) -> Result<Self> {
+        let call_kind = R::C::new(
+            command_plaintext.cmd_name(),
+            command_plaintext.runtime_params.clone(),
+        )?;
 
-        Ok(Commands {
+        Ok(CommandExecutor {
             my_account_id,
             call_kind,
-            counter: ecall_input.counter(),
+            counter: command_plaintext.counter(),
             phantom: PhantomData,
             ap: PhantomData,
         })
@@ -202,7 +263,7 @@ where
 
     fn decrypt<GK: GroupKeyOps>(ciphertext: &Ciphertext, key: &mut GK) -> Result<Option<Self>> {
         match key.decrypt(ciphertext)? {
-            Some(plaintext) => Commands::decode(&plaintext[..]).map(Some),
+            Some(plaintext) => CommandExecutor::decode(&plaintext[..]).map(Some),
             None => Ok(None),
         }
     }
